docs(davinci): mark p2-12b complete
