+++ davinci-road/plan/croquis-consumption.md
@@ -413,7 +413,7 @@ The naive lane counts raw word-boundary text matches per product name (`\.field`
 | `Croquis.provide_inject`               |       28 |   36 | `vize_croquis_cf` (26/34)                                                                                                                                                                                                                                                                                                                                                                                                                                |
 | `Croquis.race_conditions`              |        1 |    4 | `vize_croquis_cf` (1/4)                                                                                                                                                                                                                                                                                                                                                                                                                                  |
 | `Croquis.reactivity`                   |       32 |   39 | `vize_atelier_core` (0/1), `vize_davinci` (0/3), `vize_patina` (1/4)                                                                                                                                                                                                                                                                                                                                                                                     |
-| `Croquis.scopes`                       |       83 |  141 | `vize` (0/1), `vize_atelier_core` (0/4), `vize_atelier_jsx` (0/11), `vize_canon` (41/43), `vize_curator` (0/1), `vize_davinci` (0/9), `vize_patina` (4/25), `vize_s1_to_s2` (0/9)                                                                                                                                                                                                                                                                        |
+| `Croquis.scopes`                       |       83 |  142 | `vize` (0/1), `vize_atelier_core` (0/4), `vize_atelier_jsx` (0/11), `vize_canon` (41/43), `vize_curator` (0/1), `vize_davinci` (0/9), `vize_patina` (4/25), `vize_s1_to_s2` (0/10)                                                                                                                                                                                                                                                                       |
 | `Croquis.setup_context`                |        2 |    5 | `vize_croquis_cf` (1/4)                                                                                                                                                                                                                                                                                                                                                                                                                                  |
 | `Croquis.template_info`                |       18 |   20 | `vize_croquis_cf` (14/15), `vize_patina` (0/1)                                                                                                                                                                                                                                                                                                                                                                                                           |
 | `Croquis.types`                        |        9 |   29 | `vize` (0/1), `vize_atelier_sfc` (0/13), `vize_canon` (9/11), `vize_davinci` (0/3), `vize_maestro` (0/1)                                                                                                                                                                                                                                                                                                                                                 |
