+++ davinci-road/plan/phase-2-records/p2-12b.md
@@ -1,0 +1,81 @@
+# P2-12b — Fused build path + walk-count instrumentation
+
+**Source-map-free DOM builds now report the phase-2 one-walk S2 budget.
+Landed 2026-09-12 through
+[#6056](https://github.com/ubugeeei-prod/vize/pull/6056); every executable
+acceptance clause is pinned in ordinary CI suites.**
+
+## What changed
+
+`vize_s1_to_s2::pass::run_dom_transform_with_profile` is the DOM emit-specific
+product path. The full transform catalogue stays available through
+`run_transform_with_profile` for review, tests and non-DOM callers, but Vue 3
+source-map-free DOM emission no longer reports preserving fact products as
+separate pass-manager walks before codegen. `v-slot`, `v-model` and
+`hoist-static` still publish the same facts, diagnostics and provenance that
+the emitter consumes; they are folded into the selected DOM product path.
+
+Vue 2 legacy sugar is deliberately not folded. It mutates the lowered S2
+surface, so the DOM path keeps that compatibility step inside the P2-3 pass
+manager and the `BudgetObserver` still records the compatibility walk when
+Vue 2 selects it.
+
+## The measured budget
+
+The ladder now pins the source-map-free DOM build path at exactly one
+observer-facing S2 walk:
+
+| fixture       | transform walks | emit walks | build walks |
+| ------------- | --------------: | ---------: | ----------: |
+| small         |               0 |          1 |           1 |
+| medium        |               0 |          1 |           1 |
+| large         |               0 |          1 |           1 |
+| stress-deep   |               0 |          1 |           1 |
+| stress-wide   |               0 |          1 |           1 |
+| stress-interp |               0 |          1 |           1 |
+
+`crates/vize_s1_to_s2/tests/emit_budget_observer.rs` compares observed output
+against the plain emitter, asserts `total_walks == dom_walks_max`, and keeps
+emit visits below the P2-12a `[traversal.dom_*]` ceilings. `stress-wide`
+remains the control row from `budgets.toml`: attribute width is prop work, not
+stage traversal work.
+
+`crates/vize_atelier_dom/tests/davinci_s2_profile.rs` and
+`crates/vize_atelier_dom/tests/davinci_s2_build_profile.rs` pin the production
+counter surface: `davinci.s2_dom.pre_s2.walks = 0`,
+`davinci.s2_dom.transform.walks = 0`, `davinci.s2_dom.emit.walks = 1`,
+`davinci.s2_dom.total.walks = 1`, and `davinci.s2_dom.build.walks = 1` on the
+ladder.
+
+## Fusion-depth decision
+
+The open question is decided for phase 2: source-map-free DOM build may fold
+preserving semantic facts into the selected emitter product path without
+reducing diagnostic quality. The evidence is byte-identical output against the
+plain emitter and the kept `v-model` diagnostic tests under both default and
+`hoist_static: false` profiles.
+
+The boundary is also explicit. Anything that mutates S2, such as Vue 2 legacy
+sugar, remains a pass-manager step. SSR and Vapor fusion are phase-3 work, not
+a phase-2 claim.
+
+## Provenance policy
+
+Phase 2 keeps provenance fully materialized through the existing lowering/pass
+channels. No CLI ring buffer is introduced: the profiled DOM build path reaches
+the one-walk target without truncating diagnostics or provenance, and resident
+or DevTool consumers continue to read the complete materialized records.
+
+## Verification
+
+Local verification before the PR:
+
+- `cargo fmt --all --check`
+- `cargo test -p vize_s1_to_s2 --test emit_budget_observer`
+- `cargo test -p vize_atelier_dom --test davinci_s2_profile`
+- `cargo test -p vize_atelier_dom --test davinci_s2_build_profile`
+- `node --test tests/tooling/davinci-phase2-ledger.test.ts`
+
+The first local Rust run filled the disk while compiling from a cold worktree;
+after cleaning generated `target/` artifacts, the targeted suites above
+completed green.
